Agent: Extract HTTPIslandAPIClientFactory into its own module
