+++ monkey/infection_monkey/island_api_client/__init__.py
@@ -7,5 +7,6 @@ from .island_api_client_errors import (
 )
 from .i_island_api_client import IIslandAPIClient
 from .abstract_island_api_client_factory import AbstractIslandAPIClientFactory
-from .http_island_api_client import HTTPIslandAPIClient, HTTPIslandAPIClientFactory
+from .http_island_api_client import HTTPIslandAPIClient
+from .http_island_api_client_factory import HTTPIslandAPIClientFactory
 from .configuration_validator_decorator import ConfigurationValidatorDecorator
